gradioopenaicode/requirements.txt
openai
gradio
PyPDF2
python-dotenv